dedupe expand_alias
